feat: change and add logic for budgeting/expenses
Change how budget frequency is manipulated by the user - the user can
only add the monthly value for a category of budget and the yearly value
is automatically calculated. Add interaction between adding expenses and
adding budgets - if expenses for a category of budget already exist when
a budget is added, automatically subtract the total of these expenses
from both the monthly and yearly values of the budget (regardless of
frequency of expense), and vice versa if a budget already exists for a
given category when adding expenses. Add ability for a previous budget
for a category to be overwritten (no more than 1 budget per category) -
carry over expenses from the previous budget.